Divide compatibility tests (#18027)
